+++ Source/Main Dialogs/ProcessCenter/ORProcessCenter.m
@@ -313,7 +313,7 @@ SYNTHESIZE_SINGLETON_FOR_ORCLASS(ProcessCenter);
     NSString *key = [self sortColumn];
     NSArray *a = [processView tableColumns];
     NSTableColumn *column = [processView tableColumnWithIdentifier:key];
-    uint32_t i = [a count];
+    NSUInteger i = [a count];
     
     while (i-- > 0) [processView setIndicatorImage:nil inTableColumn:[a objectAtIndex:i]];
